# Refine lrn cpu forward